requirements.txt
requests>=2.31.0
pyserial>=3.5
numpy>=1.24.0
